Add switchThread synthetic atom.
Generate it whenever a command is called on a separate thread.

+++ cc/gapii/gles_types.h
@@ -533,7 +533,7 @@ typedef void* BufferDataPointer;
 
 typedef uint32_t ContextID;
 
-typedef uint32_t ThreadID;
+typedef uint64_t ThreadID;
 
 typedef int EGLBoolean;
